change format from int to Double for input parameters for proj0

+++ proj0/NBody.java
@@ -23,8 +23,8 @@ public class NBody {
     }
 
     public static void main(String[] args) {
-        double T = Integer.parseInt(args[0]);
-        double dt = Integer.parseInt(args[1]);
+        double T = Double.parseDouble(args[0]);
+        double dt = Double.parseDouble(args[1]);
         String filename = args[2];
 
         double radius = readRadius(filename);
